Scripts/requirements.txt
pillow>=8.0.0
numpy>=1.19
pandas>=1.1
matplotlib>=3.3
jupyter
notebook
ipython
